+++ path.py
@@ -470,11 +470,10 @@ class path(unicode):
         """
         if pattern is None:
             pattern = '*'
-        normcase = getattr(pattern, 'normcase', None)
         return [
             self / child
             for child in os.listdir(self)
-            if self._next_class(child).fnmatch(pattern, normcase)
+            if self._next_class(child).fnmatch(pattern)
         ]
 
     def dirs(self, pattern=None):
@@ -644,7 +643,8 @@ class path(unicode):
         """ Return ``True`` if `self.name` matches the given pattern.
 
         pattern - A filename pattern with wildcards,
-            for example ``'*.py'``.
+            for example ``'*.py'``. If the pattern contains a `normcase`
+            attribute, it is applied to the name and path prior to comparison.
 
         normcase - (optional) A function used to normalize the pattern and
             filename before matching. Defaults to self.module which defaults
@@ -652,7 +652,8 @@ class path(unicode):
 
         .. seealso:: :func:`fnmatch.fnmatch`
         """
-        normcase = normcase or self.module.normcase
+        default_normcase = getattr(pattern, 'normcase', self.module.normcase)
+        normcase = normcase or default_normcase
         name = normcase(self.name)
         pattern = normcase(pattern)
         return fnmatch.fnmatchcase(name, pattern)